requirements.txt
fastapi 
uvicorn 
pymupdf 
pytesseract 
pillow 
python-multipart 
sentence-transformers 
opencv-python-headless 
numpy
chromadb
google-generativeai
bangla
pdfplumber


fastapi 
uvicorn[standard] 
chromadb 
sentence-transformers 
pymupdf 
pdfplumber 
bangla-pdf-ocr 
python-dotenv 
groq 
opencv-python 
pillow


fastapi 
uvicorn 
python-multipart 
pymupdf 
pdfplumber 
opencv-python 
            
pillow 
bangla-pdf-ocr 
sentence-transformers 
chromadb 
pytesseract 
transformers 
torch 
indic-nlp-library 
            
google-generativeai 
layoutparser[layoutmodels]